re rganisation of code of carousel
