+++ src/App.js
@@ -4,7 +4,7 @@ import { Switch, Route } from "react-router-dom";
 import List from "./components/List";
 import Music from "./components/Music";
 import Nav from "./components/Nav";
-import Home from "./components/Home";
+import Home from "./components/homepage/Home";
 //import TrendingMusic from "./components/TrendingMusic";
 import Footer from "./components/Footer";
