+++ src/components/homepage/Home.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import TrendMusics from "./TrendMusics";
+
+import "./Home.css";
+
+const Home = () => {
+  return (
+    <>
+      <TrendMusics />
+    </>
+  );
+};
+export default Home;
